% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fname_grid.R
\name{fname_grid}
\alias{fname_grid}
\title{Get the path to the grid file}
\usage{
fname_grid(resolution)
}
\arguments{
\item{resolution}{The resolution of the grid}
}
\value{
The path to the grid file
}
\description{
This function returns the path to the grid file for a given
resolution. The grid file is a raster file that contains the
cell numbers for each pixel in the grid. These cell numbers are
used to link the data to the grid.
}
