% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resolutions.R
\name{get_resolutions}
\alias{get_resolutions}
\title{Get all available resolutions}
\usage{
get_resolutions()
}
\value{
A vector with all available resolutions
}
\description{
This function returns a vector with all available resolutions
}
