% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_cell.R
\name{grid_cell}
\alias{grid_cell}
\title{Bind data to grid cells}
\usage{
grid_cell(d, resolution)
}
\arguments{
\item{d}{The dataset with x and y coordinates}

\item{resolution}{The resolution of the grid}
}
\value{
The dataset with an additional column with the grid cell
number
}
\description{
This function needs a dataset with x and y coordinates. It expects
a tibble with columns named 'x' and 'y'. The function will add a
column 'cell' to the dataset with the grid cell number. The grid
cell number is determined by the cell number in the grid file for
the given resolution.
}
\details{
This function binds the x and y coordinates of a dataset to the
grid cell in a map with a specific resolution. The grid cell is
determined by the cell number in the grid file.
}
\examples{

data(dclutum)
d <- dclutum
grid_cell(d, 250)
}
