% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_grid.R
\name{load_grid}
\alias{load_grid}
\title{Load a grid from a file}
\usage{
load_grid(resolution)
}
\arguments{
\item{resolution}{The resolution of the grid

This package contains grid files for a set of resolutions. This
function loads the grid for a given resolution. This is loaded as a
Terra raster file.}
}
\value{
The loaded grid
}
\description{
This function loads a grid from a file. The grid is a raster file
that contains the cell numbers for each pixel in the grid. These
cell numbers are used to link the data to the grid.
}
