% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cell.R
\name{summary_cell}
\alias{summary_cell}
\title{Summarize a variable in each cell}
\usage{
summary_cell(d, variable)
}
\arguments{
\item{d}{A data frame with columns x, y, cell, and the variable to
summarize}

\item{variable}{The variable to summarize}
}
\value{
A data frame with the summary statistics]
}
\description{
This function summarizes a variable in each cell. The function
expects a data frame with columns x, y, cell, and the variable to
summarize. The function returns a data frame with the number of
observations, the 10th, 25th, 50th, 75th and 90th percentile, the
mean, the standard deviation and the median absolute deviation.
}
\details{
This function summarizes a variable in for cell.
}
